http service finalizer for automatic h2c detection
